requirements.txt
# Core ML / RAG dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-ibm>=0.1.0
langchain-chroma>=0.1.0

# IBM WatsonX SDK
ibm-watsonx-ai>=1.0.0

# Vector store
chromadb>=0.5.0

# PDF loading
pypdf>=4.0.0

# Web UI
gradio>=4.0.0

# Environment variable management
python-dotenv>=1.0.0
